backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.0
pandas==2.2.2
pyjwt==2.8.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
httpx==0.27.0
